fix(#1088): fix doc language edit when doc is validated
